start factorizing code
